fixed and refined: user services admin update issue and some others issues

+++ src/app/modules/user/user.service.ts
@@ -5,9 +5,14 @@ import ApiError from '../../../errors/ApiError'
 import { IUser } from './user.interface'
 import { User } from './user.model'
 
-const getAllUsers = async (query: { admins?: string }): Promise<IUser[]> => {
+const getAllUsers = async (
+  query: { admins?: string },
+  user: JwtPayload
+): Promise<IUser[]> => {
   const findCondition =
-    query?.admins === 'true' ? { role: { $in: ['super_admin', 'admin'] } } : {}
+    query?.admins === 'true' && user.role === 'super_admin'
+      ? { role: { $in: ['super_admin', 'admin'] } }
+      : {}
   const allUsers = await User.find(findCondition)
   return allUsers
 }
@@ -16,7 +21,7 @@ const getUser = async (email: string): Promise<IUser | null> => {
   const singleUser = await User.findOne({ email })
 
   if (!singleUser) {
-    throw new ApiError(httpStatus.NOT_FOUND, 'User not found')
+    throw new ApiError(httpStatus.NOT_FOUND, 'User not found!')
   }
 
   return singleUser
@@ -26,7 +31,7 @@ const getOwnProfile = async (user: JwtPayload): Promise<IUser | null> => {
   const profile = await User.findById(user.userId)
 
   if (!profile) {
-    throw new ApiError(httpStatus.NOT_FOUND, 'User not found')
+    throw new ApiError(httpStatus.NOT_FOUND, 'User not found!')
   }
 
   return profile
@@ -40,7 +45,7 @@ const updateUser = async (
   const targetedUser = await User.findById(id)
 
   if (!targetedUser) {
-    throw new ApiError(httpStatus.NOT_FOUND, 'User not found')
+    throw new ApiError(httpStatus.NOT_FOUND, 'User not found!')
   }
 
   if (user.role === 'user' && user.userId !== id) {
@@ -50,7 +55,11 @@ const updateUser = async (
     )
   }
 
-  if (user.role !== 'user' && targetedUser.role !== 'user') {
+  if (
+    user.role !== 'user' &&
+    targetedUser.role !== 'user' &&
+    user.userId !== targetedUser.id
+  ) {
     throw new ApiError(
       httpStatus.BAD_REQUEST,
       `You can't change other admin's profile`
@@ -62,15 +71,27 @@ const updateUser = async (
     runValidators: true,
   })
 
-  if (!updatedUser) {
-    throw new ApiError(httpStatus.NOT_FOUND, 'User not found')
-  }
-
   return updatedUser
 }
 
 const changeRole = async (email: string, payload: any): Promise<null> => {
-  const updatedUser = await User.findOneAndUpdate(
+  const targetedUser = await User.findOne({ email })
+
+  if (!targetedUser) {
+    throw new ApiError(
+      httpStatus.NOT_FOUND,
+      'User not registered with this email!'
+    )
+  }
+
+  if (targetedUser.role === 'super_admin') {
+    throw new ApiError(
+      httpStatus.BAD_REQUEST,
+      `You can't change other super admin's role!`
+    )
+  }
+
+  await User.findOneAndUpdate(
     { email },
     { role: payload.role },
     {
@@ -78,21 +99,24 @@ const changeRole = async (email: string, payload: any): Promise<null> => {
       runValidators: true,
     }
   )
-
-  if (!updatedUser) {
-    throw new ApiError(httpStatus.NOT_FOUND, 'User not found')
-  }
-
   return null
 }
 
 const deleteUser = async (id: string): Promise<null> => {
-  const user = await User.findByIdAndDelete(id)
+  const targetedUser = await User.findById(id)
 
-  if (!user) {
-    throw new ApiError(httpStatus.NOT_FOUND, 'User not found')
+  if (!targetedUser) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'User not found!')
   }
 
+  if (targetedUser.role === 'super_admin') {
+    throw new ApiError(
+      httpStatus.BAD_REQUEST,
+      `You can't delete other super admins`
+    )
+  }
+
+  await User.findByIdAndDelete(id)
   return null
 }
